Add initrd to Ubuntu test
We are hitting a race with cloud-init and udev network device renaming
when booting without an initrd. Basically cloud-init sets up the network
configuration before udev has a change to rename the device. The
configuration no longer matches the device name and networking fails.

An initrd allows udev network device renaming to happen earlier, so
this is just a bit of a band aid masking the real issue.